Move flowable logic from cf.core to cf.process
