+++ com.sap.cloud.lm.sl.cf.process/src/main/java/com/sap/cloud/lm/sl/cf/process/flowable/StartProcessAction.java
@@ -1,4 +1,4 @@
-package com.sap.cloud.lm.sl.cf.core.flowable;
+package com.sap.cloud.lm.sl.cf.process.flowable;
 
 import java.util.List;
