+++ com.sap.cloud.lm.sl.cf.web/src/main/java/com/sap/cloud/lm/sl/cf/web/api/impl/OperationsApiServiceImpl.java
@@ -29,12 +29,12 @@ import com.sap.cloud.lm.sl.cf.core.auditlogging.AuditLoggingProvider;
 import com.sap.cloud.lm.sl.cf.core.cf.CloudControllerClientProvider;
 import com.sap.cloud.lm.sl.cf.core.dao.OperationDao;
 import com.sap.cloud.lm.sl.cf.core.dao.filters.OperationFilter;
-import com.sap.cloud.lm.sl.cf.core.flowable.AbortProcessAction;
-import com.sap.cloud.lm.sl.cf.core.flowable.FlowableFacade;
-import com.sap.cloud.lm.sl.cf.core.flowable.ProcessAction;
-import com.sap.cloud.lm.sl.cf.core.flowable.ProcessActionRegistry;
-import com.sap.cloud.lm.sl.cf.core.flowable.ResumeProcessAction;
-import com.sap.cloud.lm.sl.cf.core.flowable.RetryProcessAction;
+import com.sap.cloud.lm.sl.cf.process.flowable.AbortProcessAction;
+import com.sap.cloud.lm.sl.cf.process.flowable.FlowableFacade;
+import com.sap.cloud.lm.sl.cf.process.flowable.ProcessAction;
+import com.sap.cloud.lm.sl.cf.process.flowable.ProcessActionRegistry;
+import com.sap.cloud.lm.sl.cf.process.flowable.ResumeProcessAction;
+import com.sap.cloud.lm.sl.cf.process.flowable.RetryProcessAction;
 import com.sap.cloud.lm.sl.cf.core.util.UserInfo;
 import com.sap.cloud.lm.sl.cf.persistence.message.Constants;
 import com.sap.cloud.lm.sl.cf.persistence.model.ProgressMessage;
